separated validators from descriptor for cleaner code structure
